% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qtlapi.R
\name{get_correlation_plot_data}
\alias{get_correlation_plot_data}
\title{Get the correlation.}
\usage{
get_correlation_plot_data(
  dataset,
  id,
  dataset_correlate,
  id_correlate,
  intcovar = NULL
)
}
\arguments{
\item{dataset}{the dataset identifier}

\item{id}{the identifier}

\item{intcovar}{The interactive covariate, sample.column in covar.info.}

\item{dataset.correlate}{the dataset to correlate to}

\item{id.correlate}{the identifier from the correlate dataset}
}
\value{
A named list with the data to plot
}
\description{
Get the correlation.
}
