% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qtlapi.R
\name{get_data}
\alias{get_data}
\title{Get the dataset by id (a string)}
\usage{
get_data(id, data = NULL)
}
\arguments{
\item{id}{The dataset id as a string.}

\item{data}{A string containing which data element from the dataset's data
element.}
}
\value{
The data element.
}
\description{
Get the dataset by id (a string)
}
