% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qtlapi.R
\name{get_dataset}
\alias{get_dataset}
\title{Get the dataset by id (a string)}
\usage{
get_dataset(id)
}
\arguments{
\item{id}{A string, either 'dataset.name' or just 'name'.}
}
\value{
The dataset element.
}
\description{
Get the dataset by id (a string)
}
