% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qtlapi.R
\name{get_expression}
\alias{get_expression}
\title{Get the expression data.}
\usage{
get_expression(dataset, id)
}
\arguments{
\item{dataset}{The dataset identifier.}

\item{id}{The identifier.}
}
\value{
A list with elements of tibble and list of the datatypes.
}
\description{
Get the expression data.
}
