% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qtlapi.R
\name{get_founder_coefficients}
\alias{get_founder_coefficients}
\title{Get the founder coefficients}
\usage{
get_founder_coefficients(
  dataset,
  id,
  intcovar,
  chrom,
  blup = FALSE,
  center = TRUE,
  cores = 0
)
}
\arguments{
\item{dataset}{The dataset identifier.}

\item{id}{The identifier.}

\item{intcovar}{The interactive covariate, sample.column in covar.info.}

\item{chrom}{The chromosome.}

\item{blup}{whether or not to perform BLUP}

\item{center}{whether or not to center the data}

\item{cores}{Number of cores to use (0=ALL).}
}
\value{
A nameds list with each element being a tibble with the following
        columns: id, chr, pos, and A-H
}
\description{
Get the founder coefficients
}
