% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qtlapi.R
\name{get_lod_peaks}
\alias{get_lod_peaks}
\title{Get the LOD peaks}
\usage{
get_lod_peaks(dataset, intcovar = NULL)
}
\arguments{
\item{dataset}{The dataset identifier.}

\item{intcovar}{The interactive covariate, sample.column in covar.info.}
}
\value{
a data.frame with the following columns: marker, chr, pos
and depending upons dataset$dataType the following columns:
mRNA = gene_id, symbol, gene_chrom, middle, lod
protein = protein_id, gene_id, symbol, gene_chrom, middle, lod
phenotype = data_name, short_name, description, lod
}
\description{
Get the LOD peaks
}
