% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qtlapi.R
\name{get_lod_scan}
\alias{get_lod_scan}
\title{Perform a LOD scan.}
\usage{
get_lod_scan(dataset, id, intcovar = NULL, cores = 0)
}
\arguments{
\item{dataset}{The dataset identifier.}

\item{id}{The identifier.}

\item{cores}{number of cores to use (0=ALL).}

\item{covar}{The interactive covariate.}
}
\value{
A tibble with the following columns: id, chr, pos, lod.
}
\description{
Perform a LOD scan.
}
