% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qtlapi.R
\name{get_lod_scan_by_sample}
\alias{get_lod_scan_by_sample}
\title{Perform the LOD scan for each "value" of the interactive covariate.}
\usage{
get_lod_scan_by_sample(dataset, id, intcovar, chrom, cores = 0)
}
\arguments{
\item{dataset}{The dataset identifier.}

\item{id}{The identifier.}

\item{intcovar}{The interactive covariate, sample.column in covar.info.}

\item{chrom}{The chromosome.}

\item{cores}{Number of cores to use (0=ALL).}
}
\value{
A named list with each name a "sample value" and the element is a
  tibble with the following columns: id, chr, pos, lod.
}
\description{
Perform the LOD scan for each "value" of the interactive covariate.
}
