% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qtlapi.R
\name{get_mediation}
\alias{get_mediation}
\title{Perform mediation}
\usage{
get_mediation(dataset, id, marker_id, intcovar, dataset_mediate = NULL)
}
\arguments{
\item{dataset}{The dataset identifier.}

\item{id}{The identifier.}

\item{marker_id}{The id of the marker.}

\item{intcovar}{The interactive covariate, sample.column in covar.info.}

\item{dataset_mediate}{The dataset to mediate against.}
}
\value{
A data.frame with the following columns depending on datatype:
        mRNA = gene_id, symbol, chr, pos, LOD
        protein = protein_id, gene_id, symbol, chr, pos, LOD
        phenotype = NONE
}
\description{
Perform mediation
}
