% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qtlapi.R
\name{get_snp_assoc_mapping}
\alias{get_snp_assoc_mapping}
\title{Get the SNP association mapping}
\usage{
get_snp_assoc_mapping(
  dataset,
  id,
  intcovar,
  chrom,
  location,
  window_size = 5e+05,
  cores = 0
)
}
\arguments{
\item{dataset}{The dataset identifier.}

\item{id}{The identifier.}

\item{intcovar}{The interactive covariate, sample.column in covar.info.}

\item{chrom}{The chromosome.}

\item{location}{Location on chromosome in base pairs.}

\item{window_size}{The size of the window to scan before and after location.}

\item{cores}{Number of cores to use (0=ALL).}
}
\value{
A data.frame with the following columns: snp, chr, pos, alleles,
  sdp, ensembl_gene, csq, index, interval, on_map, lod.
}
\description{
Get the SNP association mapping
}
