% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/restapi.R
\name{http_get_lodscan}
\alias{http_get_lodscan}
\title{---
get:
  description: Get scan information.
  responses:
    200:
      description: API response
      content:
        application/json:
          schema:
            type: string
            example: 5
    400:
      description: API Error
      content:
        application/json:
          schema:
            type: string
            example: Server error
---}
\usage{
http_get_lodscan(request, response)
}
\description{
---
get:
  description: Get scan information.
  responses:
    200:
      description: API response
      content:
        application/json:
          schema:
            type: string
            example: 5
    400:
      description: API Error
      content:
        application/json:
          schema:
            type: string
            example: Server error
---
}
