% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qtlapi.R
\name{is_phenotype}
\alias{is_phenotype}
\title{Check dataset to see if the datatype value is "phenotype"}
\usage{
is_phenotype(dataset)
}
\arguments{
\item{dataset}{Either list containing all values for a dataset or character
string which is the name of the dataset.}
}
\value{
TRUE if the datatype is phenotype, FALSE otherwise.
}
\description{
Check dataset to see if the datatype value is "phenotype"
}
